+++ src/app/(routes)/mover/(main)/info/edit/page.tsx
@@ -4,7 +4,7 @@ import cn from '@/utils/cn';
 import { useForm, type FieldValues, type FieldErrors } from 'react-hook-form';
 import moverEditApi from './api/moverEdit';
 import CommonButton from '@/components/common/commonBtn/commonBtn';
-import { useToaster } from '@/hooks/toaster/useToaster';
+import { useToaster } from '@/hooks/useToaster';
 
 const ul = cn('w-full', 'lg:flex lg:flex-wrap lg:w-[47%]');
 const li = cn('pt-5 pb-8 border-t border-line-100 w-full');
