+++ src/components/toast@hook/types/type.ts
@@ -1,0 +1,11 @@
+export type ToastType = 'info' | 'warn';
+export interface ToastProps {
+  type: ToastType;
+  message: string;
+  onClick: React.MouseEventHandler<HTMLDivElement>;
+}
+export interface ToastObject {
+  id: number;
+  type: ToastType;
+  message: string;
+}
