--- src/hooks/toaster/styles/toasterProvider.module.css
@@ -1,91 +1,0 @@
-.ToastContainer {
-  position: fixed;
-  display: flex;
-  flex-direction: column-reverse;
-  gap: 24px;
-  justify-content: center;
-  align-items: center;
-  top: 64px;
-  left: 50%;
-  transform: translateX(-50%);
-  pointer-events: none;
-  transition: height 0.5s ease-in-out;
-  z-index: 1000;
-}
-
-.Toast {
-  display: flex;
-  border-radius: 6px;
-  padding: 10px 16px 10px 12px;
-  justify-content: center;
-  align-items: center;
-  gap: 8px;
-  position: relative;
-  top: -8px;
-  opacity: 0;
-  transition:
-    top 0.1s ease-in-out,
-    opacity 0.1s ease-in-out;
-}
-
-.Toast.mounted {
-  top: 0;
-  opacity: 1;
-}
-
-.Icon {
-  width: 24px;
-  height: 24px;
-}
-
-.Toast.info {
-  color: #ffffff;
-  background-color: var(--primary-blue-200);
-}
-
-.Toast.warn {
-  color: #ffffff;
-  background-color: var(--secondary-red-200);
-  animation: warn 0.5s;
-}
-
-.Icon {
-  width: 24px;
-  height: 24px;
-}
-
-@keyframes warn {
-  0% {
-    transform: translateX(-1.5%);
-  }
-  10% {
-    transform: translateX(1.5%);
-  }
-  20% {
-    transform: translateX(-1.5%);
-  }
-  30% {
-    transform: translateX(1.5%);
-  }
-  40% {
-    transform: translateX(-1.5%);
-  }
-  50% {
-    transform: translateX(1.5%);
-  }
-  60% {
-    transform: translateX(-1.5%);
-  }
-  70% {
-    transform: translateX(1.5%);
-  }
-  80% {
-    transform: translateX(-1.5%);
-  }
-  90% {
-    transform: translateX(1.5%);
-  }
-  100% {
-    transform: translateX(-1.5%);
-  }
-}
